Move phasePrettyName to components/text
